Bugfix related to portals
If the platform with its riders is placed in one area and the first node is placed through a portal (the "portal group" is different) then the riders (usually) get misplaced in the void.

In normal circumstances, TryMove() will do and it's what makes the riders trigger line specials, but for cases like these it's better to treat it like a "teleport move" of sorts.