Bugfix regarding early activation and groups
Do not activate too early because it's possible there are platforms that have yet to attach themselves to our group which means not all of them would move with the origin in the first tic after activation.
This problem is apparent if the first interpolation point has a defined hold time.