Fix bug where multiple platforms use the same looping spline path
InterpolationPoint's ScanForLoop() was setting 'special1' to mark a node as already visited. And while that was fine if one platform is using a looping spline path, it's a problem if it's multiple platforms.

What was happening is those other platforms would erroneously think the spline path isn't looped because 'special1' was already set from a previous ScanForLoop() call and thus start from the next node instead.

This bug would be the same for any of GZDoom's PathFollower classes, assuming you use multiple PathFollowers on one looping spline path.