Changed PushObstacle() and SetTimeFraction() a little
Respect CANNOTPUSH and DONTTHRUST unless it's a stuck actor.
Normally, if the obstacle is pushed against a wall or solid actor etc then apply damage every 4th tic so its pain sound can be heard. But if it's not pushed against anything and 'hurtfulPush' is enabled then always apply damage. However, if there was no 'pushForce' whatsoever and 'hurtfulPush' is desired then the "damage every 4th tic" rule always applies.