This commentary is wrong
I've forgotten about the fact that during a platform's move all of its passengers are temporarily excluded from the blockmap. That means the other passenger-platform's can't find them and potentially steal them.

This contingency plan is unnecessary.

In fact, the entire thing about a platform moving another platform needs a rewrite. I'm not happy with this.