Implement tic rates for passenger and portal searching
These can be set as user variables, that is through the level editor UDB or defined as properties in sub-classes. Whichever is more convenient.

Can be used to for marginal performance gains (by not using BTIs and BLIs per tic). Or they can be used to disable creating BTIs and BLIs entirely.

Might need some tweaking and testing.