Implement convenient way to attach a special+args to platforms
Before this commit it was impossible to attach a special+args to platforms without screwing up the platform's behavior because that depended on the present state of its arguments.

Added a test sample in TEST03 with a jump-pad platform. Had to set BUMPSPECIAL on it.

This should address the need to have platforms execute specials without overhauling everything too much.

I wonder if it would be a good idea to have the FCW_Platform class have the BUMPSPECIAL flag by default for convenience's sake?