More work done on "travel speed"
Gave some descriptions in the args tooltips for UDB.

Setting travel speed can now be done from the ACS utils too.

Put more examples in SPEED. (Mainly Platform Interpolation Points and an ACS script.)

Turns out the spline bug that was fixed in the previous commit is actually useful if the points are an angle. (It aesthetically looks better imo.)

Fixed a bug where the previous point's speed would briefly be used when going to the next point.