Nevermind! Bring it back! >_<
Okay, so what can happen without this check is the following:
A platform starts with a point whose "travel time" is 0 meaning it will go to the next point in one tic. If said next point's "travel time" is non-zero then the platform would use the previous point's "super speed" instead of this point's non-zero "travel time" to reach the third point in one tic even though that isn't supposed to happen.