Make the mirror behaviour correctly mirror the new group origin
I had to use a another variable as a reference point for the mirror offsetting. And for this to work, that always had to update whenever the origin changed.
But that screwed up the orbiting behavior. So orbiters have their own reference variable because that one isn't supposed to change when the origin changes.