Don't update 'groupMirrorPos' in PostBeginPlay()
It's related to that edge case bug where a platform groups with another that's already active.

Grouping this early should never touch the group* variables concerning position, angle, pitch, and roll. (Or else this edge case bug can't be accounted for.)
But should still set the base offset for orbiters and 'groupAngle' differences; calling SetOrbitInfo() early is OK.