Just to be safe, make sure the other members are inactive when activating a group member
Yes, this is done via SetGroupOrigin(). No, it's not a good idea for this to happen in every SetGroupOrigin() call. Specifically when it's called in PlatMove() because PlatMove() is called in situations where the platform isn't supposed to be the active member of the group (eg. in MovePassengers() where the passenger is a platform.)