Create the blockmap iterators at an interval
As far as the "stress test" map is concerned, there is some minor improvement in performance (more fps) and less garbage for the GC so I'm sticking with this for now.

The only noticeable downside to this is it's no longer "in real time" which means catching things to carry might lag a little.