Make scale-affects-size optional as a user variable
This user variable is for all platform classes simply because it makes little sense to be another bitfield option. The number is getting a bit too large for my liking and for what I want to accomplish with this commit it's easier if it's a user var.
(See the commentary under BeginPlay.)

Plus, when editing generic platforms, I would like to see it among the other stuff that's for A_ChangeModel and A_SetSize.