Implement custom generic sounds
There are a couple of things I still wanted to test with this. A bug or two lingers, I think. But overall it's usable.

Now you can plop one down on the map and easily decide if and what sounds it should play. In hindsight I should have done this years ago.

This makes DemoPlatWithSnd obsolete. Adjusted the two demo map platforms. Gave TESTNastyHeart a looping sound as well because why not.