Refactor pivot behavior so it coexists with the interpolation coordinates
This fixes the bugged up test cases in PIVOTS - the ones that are orbiting around the Evil Eye decorations.
It also fundamentally changes the overall behavior with other pivot setups that are still in TEST03.
But that's more than acceptable because now the whole thing behaves closer to how I imagined it should!