Make better example of unattached pivot that rotates and follows path
These really need careful setup. I don't know if this is okay or if I need a different algorithm/logic, but trying anything else codewise hasn't worked.

With the current code it's just a fact the travel path and the recorded pivot offset can fight each other. Resulting in odd situations where with an unattached pivot that platform just rotates around it but doesn't get further away from it.