Remove "Stay On The Map" pivot argument
Shift 2 and 3 back so 1 isn't an undefined argument.

This allows me other coding possibilities that involve the pivot thing not destroying itself. Because working around that would be a pain in the ass, might as well get rid of it entirely.

And besides, idTech1 engines (or am I thinking of Hexen specifically?) already set the precedent where map spots stay on the map long after they served their purpose (because there's no harm in that) so this actor class is no different.