Implement perpetual turning
Has yet to be tested though

But if this works, it should make pivot setups more easier.
So for example you can give a platform an "active" turnspeed for roll and leave the angle changes to the interpolation paths. In other words. You don't need to adjust roll on the travel nodes themselves anymore.

You get different speeds for path following and not path following.